Fixes i2c read bug, write will require similar changes if needed for over 16 bytes. FIFO buffer size is only 16 bytes, hence flushing the buffer is required during a transfer. See src/i2c/i2c_system.c for the wait function and the src/i2c/i2c_read.c for an implementation example.

+++ src/i2c.h
@@ -63,7 +63,7 @@ uint8_t             i2c_read_reg        (  i2c_bus *i2c,
 uint8_t             *i2c_read_block     (  i2c_bus *i2c, 
                                            short addr, 
                                            short reg, 
-                                           short block_size  );
+                                           int   block_size  );
 
 /////////////////////////////////////////////////////////////
 // I2C Write ////////////////////////////////////////////////
